Split up actions component
