+++ src/UI/Pages/Agents/Components/index.ts
@@ -1,1 +1,3 @@
 export * from "./StatusLabel";
+export * from "./ActionButton";
+export * from "./KebabDropdown";
